+++ Userland/Services/WindowServer/HardwareScreenBackend.cpp
@@ -8,7 +8,7 @@
 #include "HardwareScreenBackend.h"
 #include "ScreenBackend.h"
 #include <AK/Try.h>
-#include <Kernel/API/FB.h>
+#include <Kernel/API/Graphics.h>
 #include <LibCore/System.h>
 #include <fcntl.h>
 #include <stdio.h>
